Update: Split Amimation and Skeleton builders
